framework/programs/run.py: Clear directory before starting
Currently piglit will run if a directory already exists, or die in a
fire if the results_directory is pointed at a file.

Both are bad, in the fist case if there are already X+Y tests in the
'tests' directory, and the current run only writes X tests, then Y-X
tests from the previous run will be included (which can overwrite tests
depending on the order that they were written).

This clears those out, solving the problem.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>