framework/glslparsertest: Don't add exclude extensions to the required list
glslparsertest has a syntax for the 'require_extensions' option that is
"!<extension_name>", when this is specified it means "This extensions is
not supported". There is a bug in the fast skip code in
glsl_parser_test.py that causes it to add these extensions to the fast
skip required extensions list, which is obviously wrong.

This patch adds a test for this behavior as well as fixing the behavior
to work correctly.

cc: Mark Janes <mark.a.janes@intel.com>
Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>