shader_runner: Fix get_ints on 32-bit systems.
The new ARB_vertex_attrib_64bit tests specify integer uniform values
as hex, such as 0xc21620c5.  As an integer value, this is beyond LONG_MAX
on 32-bit systems.  The intent is to parse it as an unsigned hex value and
bitcast it.

However, we still need to handle parsing values with negative signs.

Using strtoll and truncating works.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>