framework/backends/junit.py: Split _write into a separate class.
This new class is attached as an attribute to the JUnitWriter class, in
such a way that it has the same interface, and some attributes are moved
to this new class since they're only needed there. This has the
advantage of making the JUnitBackend class simpler, and makes the
writing easier to test, it will also be used in a follow up patch to
implement a different class that supports subtests.

Signed-off-by: Dylan Baker <dylanx.c.baker@intel.com>
Acked-by: Jose Fonseca <jfonseca@vmware.com> (v1)

v2: - merge the "split _sex_xml_err method" patch into this one.
      It is not possible to maintain correct behavior without doing
      this, because it would break the expected result handling.
    - Add Error message for result != pass