fbo-maxsize: Try to avoid allocating miplevels.
A 16384x16384 RGBA8888 buffer takes 1GB of RAM.  If the driver allocates
mipmap levels, that increases our storage requirements to 1.5GB.  The
test doesn't use mipmapping, so this seems like a waste.

Disabling mipmap filtering before allocating the texture provides a hint
to the driver, suggesting that it should only allocate space for the
base level.  Most Mesa drivers (Gallium and i965) follow this practice.

Signed-off-by: Kenneth Graunke <kenneth@whitecape.org>