arb_shader_image_load_store/shader-mem-barrier: Add explicit memoryBarrier to volatile subtest.
This makes sure that reads and writes of the volatile-qualified images
occur in the expected order.  It's unclear whether the GLSL volatile
qualifier was intended to have the required memory ordering
implications as it does in C.  I've bugged Khronos about it [1], but
because I haven't seen any update on the bug report for a while, this
changes the test temporarily to behave as expected regardless of what
the resolution is.  This avoids some annoying intermittent failures on
the i965 driver which doesn't provide any special memory ordering
guarantees for volatile-qualified memory.

[1] https://cvs.khronos.org/bugzilla/show_bug.cgi?id=15875