arb_gl_spirv: add GL_ACTIVE_UNIFORMS checks
Right now the main reason is to check that we are expanding uniforms
for arrays and array of arrays of structs properly. This would be even
more important with UBO/SSBO tests.

The other reason is to check that hidden uniforms are really
hidden. As far as we know, current arb_gl_spirv tests doesn't use
internally any hidden uniform, at least on the i965 driver. But we
know of cases that will be send as tests in the future.