vulkan: Prefix group names for VkRunner tests with “vulkan”
This will make it easier to distinguish tests written for Vulkan. It
makes a bit of an inconsistency because the GL and CL tests don’t have
any prefix. Ideally maybe we would add a prefix for those too, but
changing the test names at this late stage would probably cause a lot
of hassle.

Adding the prefix will avoid problems if we eventually decide to merge
the Vulkan profile into the “quick” profile because otherwise we risk
having name collisions when the same thing is tested on both Vulkan
and GL.